processing for building and floor done. to finish up by adding code to convert apiResponseItem to Building and passing it to converter or some shit

+++ FireFacilAuto/src/main/java/com/FireFacilAuto/domain/entity/building/Building.java
@@ -26,7 +26,7 @@ public class Building {
 
     public Integer overgroundFloors; //건물 내 총 지상층
 
-    public Long GFA; //연면적
+    public Double GFA; //연면적
 
     public Integer buildingClassification; //건물 주용도
